made architecture changes.
auto now has a abstract method that all autocodes extend
